change to 16 bit address

+++ eeprom.h
@@ -46,11 +46,11 @@
 *
 * uint8_t[64] data. Data to be stored. Data sits in pages in EEPROM. 
 */
-void write_single_byte(int address, uint8_t data);
-uint8_t read_single_byte(int address);
-void read_page(int address, uint8_t* temp);
+void write_single_byte(uint16_t address, uint8_t data);
+uint8_t read_single_byte(uint16_t address);
+void read_page(uint16_t address, uint8_t* temp);
 
 //this is set automatically via readbyte and readmany
-void _write_EEPROM_adr(int address);
-void _recv_EEPROM_adr(int address);
-void write_page(int address, uint8_t* data);
+void _write_EEPROM_adr(uint16_t address);
+void _recv_EEPROM_adr(uint16_t address);
+void write_page(uint16_t address, uint8_t* data);
